+++ packages/ts-components/src/components/in-article-info-card/styles.ts
@@ -117,6 +117,7 @@ export const CarouselIndicator = styled.div<{ active?: boolean }>`
   cursor: pointer;
 `;
 
+// @ts-ignore
 export const StyledCarousel = styled(ReactElasticCarousel)<{
   sectionColour: string;
 }>`