+++ packages/ts-components/src/index.ts
@@ -88,6 +88,9 @@ export {
 export {
   RecommendedFetch
 } from './components/recommended-articles/RecommendedFetch';
+export {
+  CategorisedArticles
+} from './components/categorised-articles/CategorisedArticles';
 
 // Helpers
 export { FetchProvider } from './helpers/fetch/FetchProvider';
@@ -115,6 +118,7 @@ export {
   default as safeDecodeURIComponent
 } from './utils/safeDecodeURIComponent';
 export { ArticleSidebar } from './components/article-sidebar/ArticleSidebar';
+export { ArticleAudio } from './components/article-audio/ArticleAudio';
 export {
   FeaturesCarousel
 } from './components/features-carousel/FeaturesCarousel';