+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.19.15](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.19.14...@times-components/author-profile@6.19.15) (2024-12-03)
+
+**Note:** Version bump only for package @times-components/author-profile
+
+
+
+
+
+## [6.19.14](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.19.13...@times-components/author-profile@6.19.14) (2024-12-02)
+
+**Note:** Version bump only for package @times-components/author-profile
+
+
+
+
+
 ## [6.19.13](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.19.12...@times-components/author-profile@6.19.13) (2024-11-22)
 
 **Note:** Version bump only for package @times-components/author-profile